Finish labelVisualizer style, apply style and light refactor to password components
