+++ src/components/molecules/label-visualizer/LabelVisualizer.scss
@@ -1,6 +1,21 @@
 .LabelVisualizer {
+  padding: 30px;
 
-    /*
+  .show {
+    display: flex;
+    flex-direction: row;
+
+    .label {
+      flex: 1;
+      padding-right: 20px;
+    }
+
+    .value {
+      flex: 3;
+    }
+  }
+
+  /*
     Not working. A solution could be to set an onFocus on the "edit" class element,
     however, that solution proves to be faulty, resulting in an application crash.
     
@@ -29,75 +44,78 @@
     }
     */
 
-    .editing {
-        display: flex;
-        justify-content: space-evenly;
-        .label {
-            flex: 1;
-        }
-        .textbox {
-            -moz-box-align: center;
-            align-items: center;
-            background-color: rgb(250, 251, 252);
-            border-color: rgb(223, 225, 230);
-            border-radius: 3px;
-            border-width: 2px;
-            border-style: solid;
-            box-sizing: border-box;
-            color: rgb(9, 30, 66);
-            cursor: text;
-            display: flex;
-            flex: 1 1 100%;
-            font-size: 14px;
-            -moz-box-pack: justify;
-            justify-content: space-between;
-            max-width: 100%;
-            overflow: hidden;
-            transition: background-color 0.2s ease-in-out 0s, border-color 0.2s ease-in-out 0s;
-            overflow-wrap: break-word;
-            vertical-align: top;
-            pointer-events: auto;
-            flex: 3;
-        }
-        .textbox:hover {
-            -moz-box-align: center;
-            background-color: rgb(235, 236, 240);
-            border-color: rgb(223, 225, 230);
-            border-radius: 3px;
-            border-width: 2px;
-            border-style: solid;
-            box-sizing: border-box;
-            color: rgb(9, 30, 66);
-            cursor: text;
-            display: flex;
-            flex: 1 1 100%;
-            font-size: 14px;
-            -moz-box-pack: justify;
-            justify-content: space-between;
-            max-width: 100%;
-            overflow: hidden;
-            transition: background-color 0.2s ease-in-out 0s, border-color 0.2s ease-in-out 0s;
-            overflow-wrap: break-word;
-            vertical-align: top;
-            pointer-events: auto;
-            align-items: center;
-            flex: 3;
-        }
-        .input {
-            background-color: transparent;
-            border: 0px none;
-            box-sizing: border-box;
-            color: inherit;
-            cursor: inherit;
-            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
-            font-size: 14px;
-            min-width: 0px;
-            outline: currentcolor none medium;
-            padding: 8px 6px;
-            width: 100%;
-            height: 2.57143em;
-            line-height: 1.42857;
-        }
+  .editing {
+    display: flex;
+    justify-content: space-evenly;
+    .label {
+      flex: 1;
     }
+    .textbox {
+      -moz-box-align: center;
+      align-items: center;
+      background-color: rgb(250, 251, 252);
+      border-color: rgb(223, 225, 230);
+      border-radius: 3px;
+      border-width: 2px;
+      border-style: solid;
+      box-sizing: border-box;
+      color: rgb(9, 30, 66);
+      cursor: text;
+      display: flex;
+      flex: 1 1 100%;
+      font-size: 14px;
+      -moz-box-pack: justify;
+      justify-content: space-between;
+      max-width: 100%;
+      overflow: hidden;
+      transition: background-color 0.2s ease-in-out 0s,
+        border-color 0.2s ease-in-out 0s;
+      overflow-wrap: break-word;
+      vertical-align: top;
+      pointer-events: auto;
+      flex: 3;
+    }
+    .textbox:hover {
+      -moz-box-align: center;
+      background-color: rgb(235, 236, 240);
+      border-color: rgb(223, 225, 230);
+      border-radius: 3px;
+      border-width: 2px;
+      border-style: solid;
+      box-sizing: border-box;
+      color: rgb(9, 30, 66);
+      cursor: text;
+      display: flex;
+      flex: 1 1 100%;
+      font-size: 14px;
+      -moz-box-pack: justify;
+      justify-content: space-between;
+      max-width: 100%;
+      overflow: hidden;
+      transition: background-color 0.2s ease-in-out 0s,
+        border-color 0.2s ease-in-out 0s;
+      overflow-wrap: break-word;
+      vertical-align: top;
+      pointer-events: auto;
+      align-items: center;
+      flex: 3;
+    }
+    .input {
+      background-color: transparent;
+      border: 0px none;
+      box-sizing: border-box;
+      color: inherit;
+      cursor: inherit;
+      font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto",
+        "Oxygen", "Ubuntu", "Fira Sans", "Droid Sans", "Helvetica Neue",
+        sans-serif;
+      font-size: 14px;
+      min-width: 0px;
+      outline: currentcolor none medium;
+      padding: 8px 6px;
+      width: 100%;
+      height: 2.57143em;
+      line-height: 1.42857;
+    }
+  }
 }
-
