+++ src/components/molecules/password-generator/passwordGenerator.scss
