+++ src/components/molecules/password-visualization/passwordVisualization.scss
