+++ src/components/molecules/progress-bar/progressBar.scss
