+++ src/components/templates/password/password.style.scss
@@ -41,7 +41,6 @@
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  color: #374049;
 }
 
 .rowPsw {
@@ -50,9 +49,16 @@
 }
 
 .colPsw {
-  margin-right: 15px;
   display: flex;
+  flex: 3;
   flex-direction: column;
+}
+
+.colLabel {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  padding-right: 20px;
 }
 
 .tabRow {
